+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 - Outdentation in a list caused by hard wrapping no longer closes a list.
+- Wrap `<mf-code>` with `<pre>` so that it's formatted as intended on Safari's Reader View.
 
 ## [0.0.9] - 2018-12-31